Big merge for stable release of 2.2
- Improved the code for storing the
- Fixed sort arrows not showing in the column header in vertical layout
- Restyle/changed the divider between the notes list and note textview
- Improved handling of collapsing/expanding the notes list
- Updated some deprecated methods.
- Improved support/stability on 10.9. (for example turned off
ResponsiveScrolling as the code doesn’t handle it properly yet).
- Fixed some localization stuff (menu items from English not showing up
in other languages)
- Improved 64-bit compatibility
- Fixed a lot of static analyzer and compiler warnings (see init
methods not properly setting self==[super init])

+++ ExporterManager.m
@@ -27,14 +27,14 @@
 
 - (void)awakeFromNib {
 	
-	int storageFormat = [[[GlobalPrefs defaultPrefs] notationPrefs] notesStorageFormat];
+	NSInteger storageFormat = [[[GlobalPrefs defaultPrefs] notationPrefs] notesStorageFormat];
 	[formatSelectorPopup selectItemWithTag:storageFormat];
 }
 
 - (IBAction)formatSelectorChanged:(id)sender {
 	NSSavePanel *panel = (NSSavePanel *)[sender window];
 	
-	int storageFormat = [[formatSelectorPopup selectedItem] tag];
+	NSInteger storageFormat = [[formatSelectorPopup selectedItem] tag];
 //	[panel setRequiredFileType:[NotationPrefs pathExtensionForFormat:storageFormat]];
 }
 
@@ -43,7 +43,7 @@
 	if (returnCode == NSFileHandlingPanelOKButton && notes) {
 		//write notes in chosen format
 		unsigned int i;
-		int result, storageFormat = [[formatSelectorPopup selectedItem] tag];
+		NSInteger result, storageFormat = [[formatSelectorPopup selectedItem] tag];
 		NSString *directory = nil, *filename = nil;
 		BOOL overwriteNotes = NO;
